fix: resolve compile errs resulting from error pkg rename
